Make the close-file helper a member function.
This changes the program to be C++ only.  We could make other C++
changes as well, but it is easier to see the application of C++
features through gradual evolution of the code.

The member function to close a file gives us a simple starting
point for writing the destructor, in our next revision