Change the cleanup function to a proper Destructor.
Note all of the boilerplate cleanup code which is no longer
necessary (and in fact wrong) now that C++ destructors are
responsible for resource release.  We remove this code, and
this allows the destructor to do the work of cleaning up
our resources.  Over time, further applications of C++
techniques will reduce the amount of code which is not
immediately related to the task of copying file data by
repeated reading and writing.