Update the `MemoryGuard` to represent a buffer
A buffer is not just an address to be `free`d -- it is a region
of memory with location, size, and data (contents).  The `MemoryGuard`
type now represents a reference counted region of memory and has
many useful helper functions.

The `read_buffer` function now returns a buffer, and it now relies upon
no out-parameter semantics.