Break out the FileGuard and MemoryGuard as libraries.
This program is now  somewhat representative of programming in C++.
The C++ standard library is designed with RAII in mind, allowing the
programs that need to work with resources to read more cleanly.

Notice, at this point, that the creation of useful utility libraries
with clean resource management semantics makes writing code that uses
those resources far more straightforward.  The actual main work of the
copy program remains, but the irrelevant resource management details
have been relegated to libraries.  This helps the reader and maintainer
of the code -- it is easier to see the main idea of a program when
it is not surrounded by the banalities of boilerplate resource
management.